renaming compiler/value to compiler/cvalue to avoid confusion with vm/value
